refactor segmented importation to a generic function in the `import` parent command and use it in subcommands
