+++ cmd/import/cases_segmented.go
@@ -22,21 +22,10 @@ THE SOFTWARE.
 package _import
 
 import (
-	"database/sql"
 	"errors"
-	"fmt"
-	"github.com/dustin/go-humanize"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/vbauerster/mpb/v8"
-	"github.com/vbauerster/mpb/v8/decor"
-	"golang.org/x/sync/errgroup"
-	"log"
-	"math"
-	es "snowlastic-cli/pkg/es"
 	orm "snowlastic-cli/pkg/orm"
-	"sync"
-	"time"
 )
 
 // import/casesCmd represents the import/cases command
@@ -50,127 +39,15 @@ var casesSegmentedCmd = &cobra.Command{
 		}
 		return nil
 	},
-	RunE: runCasesSegmentedImport,
+	RunE: func(cmd *cobra.Command, args []string) error {
+		var (
+			dbSchema  = "SQL_NAVEX"
+			dbTable   = "" // there are multiple, which are handled in the GetQuery return
+			indexName = "cases"
+			docType   = (&orm.Case{}).New()
+		)
+		return runSegmentedImport(dbSchema, dbTable, indexName, docType)
+	},
 }
 
 func init() {}
-
-func runCasesSegmentedImport(cmd *cobra.Command, args []string) error {
-	var (
-		dbSchema  = "SQL_NAVEX"
-		dbTable   = "" // there are multiple, which are handled in the GetQuery return
-		indexName = "cases"
-		docType   = (&orm.Case{}).New()
-
-		db    *sql.DB
-		query = docType.GetQuery(dbSchema, dbTable)
-
-		err error
-	)
-
-	db, err = generateDB(dbSchema)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer db.Close()
-
-	var segments []interface{}
-	if segmenter != "" {
-		segments, err = getSegments(db, query, quoteField(segmenter))
-	} else {
-		segments = []interface{}{"*"}
-		err = nil
-	}
-	log.Printf("processing %d segments into %s\n", len(segments), indexName)
-	if err != nil {
-		return err
-	}
-
-	g := errgroup.Group{}
-	g.SetLimit(5)
-	var wg sync.WaitGroup
-	p := mpb.New(mpb.WithWaitGroup(&wg))
-
-	for _, seg := range segments {
-		var segment interface{}
-		segment = seg
-		g.Go(func() error {
-			var thisQuery string
-			switch segment {
-			case "*", "%", "all":
-				thisQuery = query
-			case nil:
-				thisQuery = query + " WHERE " + quoteField(segmenter) + " IS NULL"
-			default:
-				thisQuery = query + " WHERE " + quoteField(segmenter) + "= " + quoteParam(segment)
-			}
-
-			var rowCount int64
-			var numBatches float64
-			rowCount, err = getRowCount(db, thisQuery)
-			if err != nil {
-				return err
-			}
-			numBatches = math.Ceil(float64(rowCount) / es.BulkInsertSize)
-
-			bar := p.AddBar(int64(numBatches),
-				//mpb.BarRemoveOnComplete(),
-				mpb.PrependDecorators(
-					//decor.Name(barName, decor.WC{W: len(barName) + 1, C: decor.DidentRight}),
-					decor.Name(fmt.Sprintf("%v", segment), decor.WCSyncSpaceR),
-					decor.CountersNoUnit("%5d/%5d", decor.WCSyncWidth),
-				),
-				mpb.AppendDecorators(
-					decor.Percentage(decor.WC{W: 5}),
-					decor.OnComplete(
-						// ETA decorator with ewma age of 30
-						decor.EwmaETA(decor.ET_STYLE_GO, 30, decor.WCSyncWidth), " done",
-					),
-				),
-			)
-
-			start := time.Now().UTC()
-			var cases = make(chan orm.SnowlasticDocument, es.BulkInsertSize)
-			h := errgroup.Group{}
-			h.SetLimit(1)
-			h.Go(func() error {
-				defer close(cases)
-				rows, err := db.Query(thisQuery)
-				if err != nil {
-					return fmt.Errorf("problem with stmt.Query(%s): %s", segment, err)
-				}
-
-				for rows.Next() {
-					var c orm.Case
-					if err := c.ScanFrom(rows); err != nil {
-						return fmt.Errorf("problem with scanning case %s: %s", c.CaseID, err)
-					}
-					cases <- &c
-				}
-				return nil
-			})
-			c, err := getElasticClient(ElasticsearchClientLocator)
-			if err != nil {
-				return fmt.Errorf("problem with getting elasticsearch clinet for segment %s: %s", segment, err)
-			}
-			batches := es.BatchEntities(cases, es.BulkInsertSize)
-			numIndexed, numErrors, err := es.BulkImportWithMPB(c, batches, indexName, bar)
-			if err != nil {
-				return fmt.Errorf("problem with bulk importing for segment %s: %s", segment, err)
-			}
-			if numErrors > 0 {
-				return errors.New(fmt.Sprintf(
-					"%s:\tIndexed [%s] documents with [%s] errors in %s (%s docs/sec)",
-					fmt.Sprintf("%s: %s=%v", indexName, segmenter, segment),
-					humanize.Comma(int64(numIndexed)),
-					humanize.Comma(int64(numErrors)),
-					time.Since(start).Truncate(time.Millisecond),
-					humanize.Comma(int64(1000.0/float64(time.Since(start)/time.Millisecond)*float64(numIndexed))),
-				))
-			}
-
-			return h.Wait()
-		})
-	}
-	return g.Wait()
-}
